deya_PP/deya_PP/requirements.txt
attrs==19.3.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
et-xmlfile==1.0.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
jdcal==1.4.1
lxml==4.5.2
numpy==1.19.0
openpyxl==3.0.4
pandas==1.0.5
parsel==1.6.0
Pillow==7.2.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
PyMySQL==0.10.0
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
queuelib==1.5.0
requests==2.24.0
Scrapy==2.2.1
service-identity==18.1.0
six==1.15.0
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.22.0
zope.interface==5.1.0
xlrd==1.2.0